move the monitor cobra command to a separate package
